Compiles!

+++ primitives/runtime/src/traits.rs
@@ -1268,9 +1268,9 @@ pub trait Benchmarking {
 	/// Type for the results of benchmarks.
 	type BenchmarkResults;
 	/// Number of steps to take between variable ranges.
-	const STEPS: usize;
+	const STEPS: u32;
 	/// Number of benchmarking repetitions.
-	const REPEATS: usize;
+	const REPEATS: u32;
 	/// Run the benchmarks for this module.
 	fn run_benchmarks() -> Vec<Self::BenchmarkResults>;
 }
